Move switch to src/main/scala-2
It's nothing but macros.
